requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
Pillow>=8.0.0
matplotlib>=3.3.0
lime>=0.2.0.1
scikit-image>=0.18.0 